Silence some warnings thrown by MSVC2015u3 while compiling as x64.

+++ gum/video/vector.hpp
@@ -41,7 +41,7 @@ struct vector : SDL_Point {
     }
 
     void normalise() {
-        int len = length();
+        int len = static_cast<int>(length());
         if(len != 0 && len != 1) {
             x /= len;
             y /= len;
